requirements.txt
# Django & Django REST Framework
Django>=5.2,<7.0
djangorestframework>=3.16,<4.0
django-cors-headers>=4.9,<5.0
django-filter>=25.2,<26.0

# WSGI Server
gunicorn>=23.0,<24.0

# State Machine
django-fsm>=3.0,<4.0

# Background Tasks
celery>=5.6,<6.0
redis>=7.2,<8.0
django-celery-beat>=2.7,<3.0

# Database
psycopg2-binary>=2.9,<3.0

# Environment & Configuration
python-dotenv>=1.2,<2.0

# Monitoring & Error Tracking
sentry-sdk>=2.53,<3.0
flower>=2.0,<3.0

# Document Generation
python-docx>=1.1,<2.0
ebooklib>=0.18,<1.0

# HTTP Requests
requests>=2.32,<3.0
httpx>=0.28,<1.0

# AI Integration
google-generativeai>=0.8,<1.0

# Payment
stripe>=12.0,<13.0

# Utilities
python-dateutil>=2.9,<3.0
Pillow>=11.0,<12.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.9,<5.0
pytest-cov>=6.0,<7.0
factory-boy>=3.3,<4.0

# Code Quality
black>=25.0,<26.0
flake8>=7.0,<8.0
isort>=5.0,<6.0
